feat:WorkRecordsResponse interface 생성
